split DataModel for LTTng
