+++ src/plugins/trace/dll/DataModel/IDataReader.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Net;
 using System.Text;
 
-namespace QuicTrace.DataModel.ETW
+namespace QuicTrace.DataModel
 {
     internal interface IDataReader
     {
