+++ src/plugins/trace/dll/DataModel/LTTng/QuicLTTngDataReader.cs
@@ -8,7 +8,7 @@ using System.Net;
 using CtfPlayback.FieldValues;
 using CtfPlayback.Metadata.Helpers;
 
-namespace QuicTrace.DataModel.ETW
+namespace QuicTrace.DataModel.LTTng
 {
     [System.Diagnostics.CodeAnalysis.SuppressMessage("Style", "IDE2001:Embedded statements must be on their own line", Justification = "<Pending>")]
     internal unsafe ref struct QuicLTTngDataReader
@@ -30,7 +30,7 @@ namespace QuicTrace.DataModel.ETW
 
         internal IntegerLiteral Preprocess()
         {
-            var val = (CtfIntegerValue)Payload.FieldsByName[String.Concat("_arg", Index.ToString())];
+            var val = (CtfIntegerValue)Payload.FieldsByName[string.Concat("_arg", Index.ToString())];
             Index++;
             return val.Value;
         }
@@ -73,31 +73,31 @@ namespace QuicTrace.DataModel.ETW
 
         internal byte[] ReadBytes()
         {
-            var argstr = String.Concat("_arg", Index.ToString());
+            var argstr = string.Concat("_arg", Index.ToString());
             Index++;
             return ((CtfArrayValue)Payload.FieldsByName[argstr]).ReadAsUInt8Array();
         }
 
         internal string ReadString()
         {
-            var val = ((CtfStringValue)Payload.FieldsByName[String.Concat("_arg", Index.ToString())]).Value;
+            var val = ((CtfStringValue)Payload.FieldsByName[string.Concat("_arg", Index.ToString())]).Value;
             Index++;
             return val;
         }
 
         internal IPEndPoint ReadAddress()
         {
-            var argstr = String.Concat("_arg", Index.ToString());
+            var argstr = string.Concat("_arg", Index.ToString());
             Index++;
-            var length = UInt32.Parse(Payload.FieldsByName[String.Concat(argstr, "_len")].GetValueAsString());
+            var length = uint.Parse(Payload.FieldsByName[string.Concat(argstr, "_len")].GetValueAsString());
             if (length == 0)
             {
                 return new IPEndPoint(IPAddress.None, 0);
             }
 
             var buf = ((CtfArrayValue)Payload.FieldsByName[argstr]).ReadAsUInt8Array();
-            int family = buf[0] | ((ushort)buf[1] << 8);
-            int port = (ushort)buf[3] | ((ushort)buf[2] << 8);
+            int family = buf[0] | buf[1] << 8;
+            int port = buf[3] | buf[2] << 8;
 
             if (family == 0) // unspecified
             {
